pyproject.toml
[project]
name = "call_your_loved_ones"
version = "0.1.0"
description = ""
authors = [{ name = "mrkaye97", email = "mrkaye97@gmail.com" }]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard] (>=0.121.3,<0.122.0)",
    "asyncpg>=0.29.0",
    "pyjwt>=2.8.0",
    "bcrypt>=4.0.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "isort>=5.13.0",
    "black>=24.0.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 105

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = []

[tool.black]
line-length = 105

[tool.isort]
profile = "black"
line_length = 105

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
follow_imports = "normal"
ignore_missing_imports = false
check_untyped_defs = true

# Per-module options for third-party libraries without stubs
[[tool.mypy.overrides]]
module = ["fastapi.*", "asyncpg.*", "bcrypt.*", "jwt.*"]
ignore_missing_imports = true
